More MSVC-related fixes: - "implicit" narrowing is now explicit (IMO `uint32_t{some_u64}` is no less explicit than `static_cast<uint32_t>(some_u64)`, but I didn't write the compiler). - `CastSpellBuilder` is a struct, not a class.
Annoyingly, since node.exe/node.lib has statically linked zlib, I've got to figure something else out for that.

+++ src/ChatCommandBuilderBuilder.cpp
@@ -155,9 +155,9 @@ ChatCommandBuilder ChatCommandBuilderBuilder::build() {
 			throw IncompleteVariantError();
 		}
 		if (help) {
-			return {m_name.c_str(), *dispatch_handler(std::get<size_t>(*handler)), *help, *security_level, *allow_console};
+			return {m_name.c_str(), *dispatch_handler(std::get<size_t>(*handler)), *help, static_cast<uint32_t>(*security_level), *allow_console};
 		}
-		return {m_name.c_str(), *dispatch_handler(std::get<size_t>(*handler)), *security_level, *allow_console};
+		return {m_name.c_str(), *dispatch_handler(std::get<size_t>(*handler)), static_cast<uint32_t>(*security_level), *allow_console};
 	}
 	throw IncompleteVariantError();
 }
